requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
dataclasses==0.5
idna==2.7
lxml==4.2.4
mir.dlsite==0.5.0
requests==2.19.1
urllib3==1.23
